requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gunicorn==21.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1